+++ CHANGES.md
@@ -5,7 +5,7 @@ Here you can see the full list of changes between each PyScript release.
 
 Version 0.2.1
 -----------
-- added mockSetup() and windowProxy to router module
+- added mockSetup() and proxy (for window) to router module
 - mockSetup() prevents page reload for __Jasmine 2.0__ tests
 
 Version 0.2.0